requirements.txt
dlib-binary==19.21.99
Keras==2.4.3
matplotlib==3.3.3
numpy==1.18.5
opencv-contrib-python==4.2.0.34
opencv-python==3.4.2.17
scikit-learn==0.23.2
scipy==1.5.4
tensorflow==2.3.1